Fix bug with ID
Issue was discovered where editting a task caused other task details to be editted.
Debug revealed that the corrent ID was not loaded from local storage and
this resulted in multiple tasks with the same ID. As the logic loops task
ID's while events are detected, this caused multiple tasks to be updated.
Fix for the issue has been to load the correct ID from local storage when
tasks are rendered to the page.